# api/api.yaml
---
openapi: 3.0.3
info:
  title: "Artifact Store API specification"
  version: "0.1.0"
paths:
  /v1/repositories:
    get:
      summary: "Get all available repositories"
      operationId: "getRepositories"
      responses:
        '200':
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                $ref: '#/components/schemas/Repository'
    post:
      summary: "Create repository"
      operationId: "addRepository"
      operation: addRepository
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                repository:
                  type: string
                artifact:
                  type: string
                  description: "The artifact kind the repository will host"
                  enum:
                    - helm
              required:
                - name
                - repository
                - artifact
      responses:
        '201':
          description: "Repository created"
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                required:
                  - name
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: "Conflict — repository already exists"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/helm:
    get:
      summary: "Get all available Helm charts"
      operationId: "getCharts"
      responses:
        '200':
          description: "List of Helm charts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chart'
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/helm/{name}/versions:
    get:
      summary: "Get list of versions for a Helm chart"
      operationId: "getChartVersions"
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: "Helm chart name"
      responses:
        '200':
          description: "Helm chart version list"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/helm/{repository}/{name}/{version}:
    get:
      summary: "Download version of Helm chart"
      operationId: "getChart"
      parameters:
        - name: repository
          in: path
          required: false
          schema:
            type: string
            default: default
          description: "Chart repository"
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: "Helm chart name"
        - name: version # TODO: make optional, default to 'latest'
          in: path
          required: true
          schema:
            type: string
          description: "Helm chart version"
      responses:
        '200':
          description: "Helm chart download"
          content:
            application/gzip:
              schema:
                type: string
                format: binary
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/helm/:
    post:
      summary: "Add Helm chart"
      operationId: "addChart"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                repository:
                  type: string
                chart:
                  type: string
                  format: binary
                  description: "The packaged chart file (.tgz)"
      responses:
        '201':
          description: "Helm chart created"
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  version:
                    type: string
                required:
                  - name
                  - version
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: "Conflict — chart version already exists"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
    Repository:
      type: object
      properties:
        name:
          type: string
        artifact:
          type: string
          enum:
            - chart
    Chart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string