# internal/api/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for RepositoryArtifact.
const (
	RepositoryArtifactChart RepositoryArtifact = "chart"
)

// Valid indicates whether the value is a known member of the RepositoryArtifact enum.
func (e RepositoryArtifact) Valid() bool {
	switch e {
	case RepositoryArtifactChart:
		return true
	default:
		return false
	}
}

// Defines values for AddRepositoryMultipartBodyArtifact.
const (
	Helm AddRepositoryMultipartBodyArtifact = "helm"
)

// Valid indicates whether the value is a known member of the AddRepositoryMultipartBodyArtifact enum.
func (e AddRepositoryMultipartBodyArtifact) Valid() bool {
	switch e {
	case Helm:
		return true
	default:
		return false
	}
}

// Chart defines model for Chart.
type Chart struct {
	Id   *int64  `json:"id,omitempty"`
	Name *string `json:"name,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Code    *int    `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// Repository defines model for Repository.
type Repository struct {
	Artifact *RepositoryArtifact `json:"artifact,omitempty"`
	Name     *string             `json:"name,omitempty"`
}

// RepositoryArtifact defines model for Repository.Artifact.
type RepositoryArtifact string

// AddChartMultipartBody defines parameters for AddChart.
type AddChartMultipartBody struct {
	// Chart The packaged chart file (.tgz)
	Chart      *openapi_types.File `json:"chart,omitempty"`
	Name       *string             `json:"name,omitempty"`
	Repository *string             `json:"repository,omitempty"`
}

// AddRepositoryMultipartBody defines parameters for AddRepository.
type AddRepositoryMultipartBody struct {
	// Artifact The artifact kind the repository will host
	Artifact   AddRepositoryMultipartBodyArtifact `json:"artifact"`
	Name       string                             `json:"name"`
	Repository string                             `json:"repository"`
}

// AddRepositoryMultipartBodyArtifact defines parameters for AddRepository.
type AddRepositoryMultipartBodyArtifact string

// AddChartMultipartRequestBody defines body for AddChart for multipart/form-data ContentType.
type AddChartMultipartRequestBody AddChartMultipartBody

// AddRepositoryMultipartRequestBody defines body for AddRepository for multipart/form-data ContentType.
type AddRepositoryMultipartRequestBody AddRepositoryMultipartBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get all available Helm charts
	// (GET /v1/helm)
	GetCharts(w http.ResponseWriter, r *http.Request)
	// Add Helm chart
	// (POST /v1/helm/)
	AddChart(w http.ResponseWriter, r *http.Request)
	// Get list of versions for a Helm chart
	// (GET /v1/helm/{name}/versions)
	GetChartVersions(w http.ResponseWriter, r *http.Request, name string)
	// Download version of Helm chart
	// (GET /v1/helm/{repository}/{name}/{version})
	GetChart(w http.ResponseWriter, r *http.Request, repository string, name string, version string)
	// Get all available repositories
	// (GET /v1/repositories)
	GetRepositories(w http.ResponseWriter, r *http.Request)
	// Create repository
	// (POST /v1/repositories)
	AddRepository(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetCharts operation middleware
func (siw *ServerInterfaceWrapper) GetCharts(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetCharts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddChart operation middleware
func (siw *ServerInterfaceWrapper) AddChart(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddChart(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetChartVersions operation middleware
func (siw *ServerInterfaceWrapper) GetChartVersions(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetChartVersions(w, r, name)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetChart operation middleware
func (siw *ServerInterfaceWrapper) GetChart(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "repository" -------------
	var repository string

	err = runtime.BindStyledParameterWithOptions("simple", "repository", r.PathValue("repository"), &repository, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: false, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "repository", Err: err})
		return
	}

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", r.PathValue("name"), &name, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "name", Err: err})
		return
	}

	// ------------- Path parameter "version" -------------
	var version string

	err = runtime.BindStyledParameterWithOptions("simple", "version", r.PathValue("version"), &version, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "version", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetChart(w, r, repository, name, version)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRepositories operation middleware
func (siw *ServerInterfaceWrapper) GetRepositories(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRepositories(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddRepository operation middleware
func (siw *ServerInterfaceWrapper) AddRepository(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddRepository(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/v1/helm", wrapper.GetCharts)
	m.HandleFunc("POST "+options.BaseURL+"/v1/helm/", wrapper.AddChart)
	m.HandleFunc("GET "+options.BaseURL+"/v1/helm/{name}/versions", wrapper.GetChartVersions)
	m.HandleFunc("GET "+options.BaseURL+"/v1/helm/{repository}/{name}/{version}", wrapper.GetChart)
	m.HandleFunc("GET "+options.BaseURL+"/v1/repositories", wrapper.GetRepositories)
	m.HandleFunc("POST "+options.BaseURL+"/v1/repositories", wrapper.AddRepository)

	return m
}